More work on count and replace block commands
Commands classes are now situated in their own files for clarity.

Replace blocks now correctly sets the block state an the lighting issues are gone.

Initial work on the translated output from the commands.
