+++ build.gradle
@@ -46,6 +46,8 @@ configurations {
 minecraft {
 	mappings channel: 'snapshot', version: "${config.mcp_mappings}"
 
+	accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
+
 	runs {
 		client {
 			workingDirectory project.file('run')
@@ -80,16 +82,16 @@ dependencies {
 	shade "com.typesafe:config:1.2.1"
 }
 
-//processResources {
+processResources {
 //	inputs.property "version", project.version
 //	inputs.property "mc_version", project.config.mc_version
-//
+
 //	from(sourceSets.main.resources.srcDirs) {
 //		include 'mcmod.info'
 //		expand 'version': project.version, 'mc_version': project.config.mc_version
 //	}
-//	rename '(.+_at.cfg)', 'META-INF/$1'
-//}
+	rename '(.+_at.cfg)', 'META-INF/$1'
+}
 
 version = "${config.mc_version}-${config.mod_version}"
